LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/flow-content-0/dialog-expected.txt




PASS Closed dialog in width: 540px iframe
PASS Open dialog in width: 540px iframe
FAIL Modal dialog in width: 540px iframe assert_equals: top expected "56px" but got "8px"
PASS Closed dialog in width: 538px iframe
PASS Open dialog in width: 538px iframe
FAIL Modal dialog in width: 538px iframe assert_equals: top expected "56px" but got "8px"
